Add message callback for NameText:SetTextColor
I maintain a KNP plugin that extends the options to override health bar color by auras or by target name.

This helps players see which targets are missing a debuff, or flagging priority targets in mythic+/raid.

For tank mode, I'd like to override the NameText instead of the HealthBar, since the Tank Mode colors are super useful.

This change adds a new message when changing the name text so that a plugin can override the defaults.